+++ src/main/java/fit/health/fithealthapi/controllers/RecipeController.java
@@ -3,15 +3,17 @@ package fit.health.fithealthapi.controllers;
 import fit.health.fithealthapi.exceptions.IngredientNotFoundException;
 import fit.health.fithealthapi.exceptions.RecipeNotFoundException;
 import fit.health.fithealthapi.model.Recipe;
+import fit.health.fithealthapi.model.User;
 import fit.health.fithealthapi.model.dto.RecipeSearchRequest;
-import fit.health.fithealthapi.model.enums.Allergen;
-import fit.health.fithealthapi.model.enums.DietaryPreference;
-import fit.health.fithealthapi.model.enums.HealthConditionSuitability;
+import fit.health.fithealthapi.model.enums.*;
 import fit.health.fithealthapi.services.RecipeService;
 import fit.health.fithealthapi.services.SharedService;
+import fit.health.fithealthapi.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,13 +25,23 @@ public class RecipeController {
     @Autowired
     private RecipeService recipeService;
     @Autowired
+    private UserService userService;
+    @Autowired
     private SharedService sharedService;
 
     /**
      * Create a new recipe.
      */
     @PostMapping
-    public ResponseEntity<Recipe> createRecipe(@RequestBody Recipe recipe) {
+    public ResponseEntity<?> createRecipe(@RequestBody Recipe recipe) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        User owner = userService.getUserByUsername(username);
+        if (owner == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid user");
+        }
+
+        recipe.setOwner(owner);
         Recipe createdRecipe = recipeService.saveRecipe(recipe);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdRecipe);
     }
@@ -62,8 +74,13 @@ public class RecipeController {
      * Update an existing recipe.
      */
     @PutMapping("/{id}")
-    public ResponseEntity<Recipe> updateRecipe(@PathVariable Long id, @RequestBody Recipe updatedRecipe) {
+    public ResponseEntity<?> updateRecipe(@PathVariable Long id, @RequestBody Recipe updatedRecipe, Authentication authentication) {
         try {
+            String currentUsername = authentication.getName();
+            User currentUser = userService.getUserByUsername(currentUsername);
+            if (!currentUser.getRole().equals(Role.ADMIN) && !currentUser.getId().equals(recipeService.getRecipeById(id).getOwner().getId())) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You don't have permission to update this meal plan.");
+            }
             Recipe recipe = recipeService.updateRecipe(id, updatedRecipe);
             return ResponseEntity.ok(recipe);
         } catch (RecipeNotFoundException e) {
@@ -79,8 +96,13 @@ public class RecipeController {
      * Delete a recipe by ID.
      */
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteRecipe(@PathVariable Long id) {
+    public ResponseEntity<?> deleteRecipe(@PathVariable Long id, Authentication authentication) {
         try {
+            String currentUsername = authentication.getName();
+            User currentUser = userService.getUserByUsername(currentUsername);
+            if (!currentUser.getRole().equals(Role.ADMIN) && !currentUser.getId().equals(recipeService.getRecipeById(id).getOwner().getId())) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You don't have permission to delete this meal plan.");
+            }
             recipeService.deleteRecipe(id);
             return ResponseEntity.noContent().build();
         } catch (RecipeNotFoundException e) {
@@ -92,19 +114,10 @@ public class RecipeController {
     public ResponseEntity<List<Recipe>> searchRecipes(
             @RequestBody RecipeSearchRequest searchRequest
     ) {
-        List<DietaryPreference> dietaryPreferences = sharedService.convertToDietaryPreferences(searchRequest.getDietaryPreferences());
-        List<Allergen> allergens = sharedService.convertToAllergens(searchRequest.getAllergens());
-        List<HealthConditionSuitability> healthConditions = sharedService.convertToHealthConditionSuitability(searchRequest.getHealthConditions());
 
+        searchRequest.setConditionSuitability(sharedService.convertToHealthConditionSuitability(searchRequest.getHealthConditions()));
         List<Recipe> recipes = recipeService.searchRecipes(
-                dietaryPreferences,
-                allergens,
-                healthConditions,
-                searchRequest.getIngredientNames(),
-                searchRequest.getMinCalories(),
-                searchRequest.getMaxCalories(),
-                searchRequest.getMaxTotalTime(),
-                searchRequest.getName()
+              searchRequest
         );
         return ResponseEntity.ok(recipes);
     }
